Split Components and ComponentGroups into two controllers
