+++ app/Http/Controllers/Dashboard/ComponentController.php
@@ -15,9 +15,6 @@ use AltThree\Validator\ValidationException;
 use CachetHQ\Cachet\Bus\Commands\Component\CreateComponentCommand;
 use CachetHQ\Cachet\Bus\Commands\Component\RemoveComponentCommand;
 use CachetHQ\Cachet\Bus\Commands\Component\UpdateComponentCommand;
-use CachetHQ\Cachet\Bus\Commands\ComponentGroup\CreateComponentGroupCommand;
-use CachetHQ\Cachet\Bus\Commands\ComponentGroup\RemoveComponentGroupCommand;
-use CachetHQ\Cachet\Bus\Commands\ComponentGroup\UpdateComponentGroupCommand;
 use CachetHQ\Cachet\Models\Component;
 use CachetHQ\Cachet\Models\ComponentGroup;
 use CachetHQ\Cachet\Models\Tag;
@@ -25,6 +22,11 @@ use GrahamCampbell\Binput\Facades\Binput;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\View;
 
+/**
+ * This is the component controller class.
+ *
+ * @author James Brooks <james@alt-three.com>
+ */
 class ComponentController extends Controller
 {
     /**
@@ -76,21 +78,6 @@ class ComponentController extends Controller
         return View::make('dashboard.components.index')
             ->withPageTitle(trans_choice('dashboard.components.components', 2).' - '.trans('dashboard.dashboard'))
             ->withComponents($components)
-            ->withSubMenu($this->subMenu);
-    }
-
-    /**
-     * Shows the component groups view.
-     *
-     * @return \Illuminate\View\View
-     */
-    public function showComponentGroups()
-    {
-        $this->subMenu['groups']['active'] = true;
-
-        return View::make('dashboard.components.groups.index')
-            ->withPageTitle(trans_choice('dashboard.components.groups.groups', 2).' - '.trans('dashboard.dashboard'))
-            ->withGroups(ComponentGroup::orderBy('order')->get())
             ->withSubMenu($this->subMenu);
     }
 
@@ -226,98 +213,5 @@ class ComponentController extends Controller
 
         return cachet_redirect('dashboard.components')
             ->withSuccess(sprintf('%s %s', trans('dashboard.notifications.awesome'), trans('dashboard.components.delete.success')));
-    }
-
-    /**
-     * Deletes a given component group.
-     *
-     * @param \CachetHQ\Cachet\Models\ComponentGroup $group
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function deleteComponentGroupAction(ComponentGroup $group)
-    {
-        dispatch(new RemoveComponentGroupCommand($group));
-
-        return cachet_redirect('dashboard.components.groups')
-            ->withSuccess(sprintf('%s %s', trans('dashboard.notifications.awesome'), trans('dashboard.components.delete.success')));
-    }
-
-    /**
-     * Shows the add component group view.
-     *
-     * @return \Illuminate\View\View
-     */
-    public function showAddComponentGroup()
-    {
-        return View::make('dashboard.components.groups.add')
-            ->withPageTitle(trans('dashboard.components.groups.add.title').' - '.trans('dashboard.dashboard'));
-    }
-
-    /**
-     * Shows the edit component group view.
-     *
-     * @param \CachetHQ\Cachet\Models\ComponentGroup $group
-     *
-     * @return \Illuminate\View\View
-     */
-    public function showEditComponentGroup(ComponentGroup $group)
-    {
-        return View::make('dashboard.components.groups.edit')
-            ->withPageTitle(trans('dashboard.components.groups.edit.title').' - '.trans('dashboard.dashboard'))
-            ->withGroup($group);
-    }
-
-    /**
-     * Creates a new component.
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function postAddComponentGroup()
-    {
-        try {
-            $group = dispatch(new CreateComponentGroupCommand(
-                Binput::get('name'),
-                Binput::get('order', 0),
-                Binput::get('collapsed'),
-                Binput::get('visible')
-            ));
-        } catch (ValidationException $e) {
-            return cachet_redirect('dashboard.components.groups.create')
-                ->withInput(Binput::all())
-                ->withTitle(sprintf('%s %s', trans('dashboard.notifications.whoops'), trans('dashboard.components.groups.add.failure')))
-                ->withErrors($e->getMessageBag());
-        }
-
-        return cachet_redirect('dashboard.components.groups')
-            ->withSuccess(sprintf('%s %s', trans('dashboard.notifications.awesome'), trans('dashboard.components.groups.add.success')));
-    }
-
-    /**
-     * Updates a component group.
-     *
-     * @param \CachetHQ\Cachet\Models\ComponentGroup $group
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function updateComponentGroupAction(ComponentGroup $group)
-    {
-        try {
-            $group = dispatch(new UpdateComponentGroupCommand(
-                $group,
-                Binput::get('name'),
-                $group->order,
-                Binput::get('collapsed'),
-                Binput::get('visible')
-            ));
-        } catch (ValidationException $e) {
-            return cachet_redirect('dashboard.components.groups.edit', [$group->id])
-                ->withInput(Binput::all())
-                ->withTitle(sprintf('%s %s', trans('dashboard.notifications.whoops'), trans('dashboard.components.groups.edit.failure')))
-                ->withErrors($e->getMessageBag());
-        }
-
-        return cachet_redirect('dashboard.components.groups.edit', [$group->id])
-            ->withSuccess(sprintf('%s %s', trans('dashboard.notifications.awesome'), trans('dashboard.components.groups.edit.success')));
     }
 }
